gucci

+++ packages/hardhat/contracts/RocketFactory.sol
@@ -3,73 +3,96 @@ pragma solidity >=0.4.16 <0.9.0;
 pragma experimental ABIEncoderV2;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
-
 contract RocketFactory {
-
     Transaction[] transactionsList;
     //mapping for
 
-    struct  Transaction{
-        uint id;
+    struct Transaction {
+        uint256 id;
         address owner;
         address receiver;
-        uint  deadline;
-        address  ERC20TokenAddress;
-        uint  amount;
-        uint  tip;    
+        uint256 deadline;
+        address ERC20TokenAddress;
+        uint256 amount;
+        uint256 tip;
         bool pending;
     }
 
     //creates a new transaction for later scheduling
-    function createTransaction(address receiver, address ERC20TokenAddress, uint amount, uint deadline, uint tip) public {
+    function createTransaction(
+        address receiver,
+        address ERC20TokenAddress,
+        uint256 amount,
+        uint256 deadline,
+        uint256 tip
+    ) public {
         // Check if they have the ERC20Token amount in their wallet
-        uint dueBy = block.timestamp + deadline; //in seconds
+        uint256 dueBy = block.timestamp + deadline; //in seconds
         IERC20 token = IERC20(ERC20TokenAddress);
         require(
             token.allowance(msg.sender, address(this)) >= amount,
             "Not enough tokens to complete tx"
-            );
-            
+        );
+
         //Send that amount to the contract
         bool sent = token.transferFrom(msg.sender, address(this), amount);
         require(sent, "transfer fialed");
 
-        Transaction memory newTransaction = Transaction(transactionsList.length ,msg.sender, receiver, dueBy, ERC20TokenAddress, amount, tip, true);
+        Transaction memory newTransaction = Transaction(
+            transactionsList.length,
+            msg.sender,
+            receiver,
+            dueBy,
+            ERC20TokenAddress,
+            amount,
+            tip,
+            true
+        );
         transactionsList.push(newTransaction);
-    
     }
 
     //TODO
     // 1. send the money from the escrow account to the receiver
     // 2. (later) add a modifier that only allows approved liquidators to execute transaction
     //who can execute this transaction? only the liquidator bots
-    function executeTransaction(uint id) public returns (Transaction memory) {
+    function executeTransaction(uint256 id)
+        public
+        returns (Transaction memory)
+    {
         //ensure that the transaction is not pending
         Transaction memory scheduledTx = getTransaction(id);
         require(scheduledTx.pending == true, "Already executed");
         // require( block.timestamp >= scheduledTx.deadline,"Not ready to be executed");
-        
+
         //take the money from the lending pool and send it to the receiver
         //do this later
 
-
         // create a new struct and update the transactionsList
-        Transaction memory newTransaction = Transaction(id,scheduledTx.owner, scheduledTx.receiver, scheduledTx.deadline, scheduledTx.ERC20TokenAddress, scheduledTx.amount, scheduledTx.tip, false);
+        Transaction memory newTransaction = Transaction(
+            id,
+            scheduledTx.owner,
+            scheduledTx.receiver,
+            scheduledTx.deadline,
+            scheduledTx.ERC20TokenAddress,
+            scheduledTx.amount,
+            scheduledTx.tip,
+            false
+        );
         transactionsList[id] = newTransaction;
         return newTransaction;
-
-
     }
 
     //get the current transactions in the array
-    function getAllTransactions() public view returns (Transaction[] memory ) {
+    function getAllTransactions() public view returns (Transaction[] memory) {
         return transactionsList;
     }
 
-    function getTransaction(uint id) public view returns (Transaction memory) {
+    function getTransaction(uint256 id)
+        public
+        view
+        returns (Transaction memory)
+    {
         require(id <= transactionsList.length);
         return transactionsList[id];
-        
-        
     }
 }
